larynx-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
cssselect==1.3.0
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.14
fuzzywuzzy==0.18.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
jiter==0.10.0
joblib==1.5.1
Levenshtein==0.27.1
lxml==6.0.0
more-itertools==10.7.0
nltk==3.9.1
numpy==2.3.1
openai==1.95.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
postgrest==1.1.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
RapidFuzz==3.13.0
realtime==2.5.3
regex==2024.11.6
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
talon==1.4.4 #THIS WILL CAUSE AN ERROR install chardet and not cchardet and then install talon without cchardet
threadpoolctl==3.6.0
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.35.0
webencodings==0.5.1
websockets==15.0.1
